Move declaration of matplotlib backend to unit tests
This fixes a bug where python unittest discover on Linux yielded this
error:

TclError: couldn't connect to display "localhost:...."

Unit tests arre inherently non-interactive, and in matplotlib, your have
to explicitly turn interactive off. Before that was done inside the
wnd-charm library, but the place to declare matplotlib backend is the
point of usage, a.k.a. the unit tests themselves.