2x speedup for time to load sigfiles
The call to zip to unpack the feature names and values from a sig file
was an absolute dog. Now, if we know how many featues are in the file
ahead of time, be more efficient about unpacking the data to immediately
go exactly where they're gonna end up anyway. Fall back to old method as
a last resort if number of feature in a file is not know ahead of time.